pyproject.toml
[project]
name = "metapy"
version = "1.0.0"
description = "A real-world customer support simulation environment"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "openai>=1.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
server = "app:main"
